# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Backlog Syntax API",
    "version": "0.1.0",
    "description": "Contract-first REST API for a secure, multi-tenant, agent-native backlog."
  },
  "servers": [
    {
      "url": "http://localhost:8787"
    }
  ],
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-backlog_session",
        "description": "Better Auth browser session cookie. The configured name may differ locally."
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "PAT or OAuth 2.1 access token"
      }
    },
    "schemas": {
      "WorkspaceContext": {
        "type": "object",
        "properties": {
          "workspace": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid",
                "example": "019641a8-8c54-7f6c-8d2f-3fd1eb8b7531"
              },
              "name": {
                "type": "string",
                "minLength": 1,
                "maxLength": 120
              },
              "slug": {
                "type": "string",
                "minLength": 1,
                "maxLength": 80
              }
            },
            "required": [
              "id",
              "name",
              "slug"
            ],
            "additionalProperties": false
          },
          "principal": {
            "type": "object",
            "properties": {
              "subjectType": {
                "type": "string",
                "enum": [
                  "user",
                  "service_account"
                ]
              },
              "subjectId": {
                "type": "string",
                "format": "uuid",
                "example": "019641a8-8c54-7f6c-8d2f-3fd1eb8b7531"
              },
              "role": {
                "type": "string",
                "enum": [
                  "owner",
                  "admin",
                  "member",
                  "viewer"
                ]
              }
            },
            "required": [
              "subjectType",
              "subjectId",
              "role"
            ],
            "additionalProperties": false
          },
          "projects": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "format": "uuid",
                  "example": "019641a8-8c54-7f6c-8d2f-3fd1eb8b7531"
                },
                "name": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 120
                },
                "slug": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 80
                },
                "updatedAt": {
                  "type": "string",
                  "format": "date-time",
                  "example": "2026-08-31T12:00:00Z"
                }
              },
              "required": [
                "id",
                "name",
                "slug",
                "updatedAt"
              ],
              "additionalProperties": false
            }
          }
        },
        "required": [
          "workspace",
          "principal",
          "projects"
        ],
        "additionalProperties": false
      },
      "ProblemDetail": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "format": "uri"
          },
          "title": {
            "type": "string",
            "minLength": 1
          },
          "status": {
            "type": "integer",
            "minimum": 400,
            "maximum": 599
          },
          "detail": {
            "type": "string",
            "minLength": 1
          },
          "instance": {
            "type": "string",
            "minLength": 1
          },
          "traceId": {
            "type": "string",
            "minLength": 1
          },
          "code": {
            "type": "string",
            "enum": [
              "invalid_request",
              "unauthenticated",
              "forbidden",
              "not_found",
              "conflict",
              "stale_version",
              "rate_limited",
              "payload_too_large",
              "internal_error",
              "dependency_unavailable",
              "reauthentication_required",
              "sole_owner_workspace",
              "invalid_confirmation"
            ]
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "field": {
                  "type": "string",
                  "minLength": 1
                },
                "message": {
                  "type": "string",
                  "minLength": 1
                }
              },
              "required": [
                "field",
                "message"
              ],
              "additionalProperties": false
            }
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "traceId",
          "code"
        ],
        "additionalProperties": false
      },
...
    },
...
  },
  "paths": {
    "/v1/workspaces/{workspaceId}/context": {
      "get": {
        "operationId": "getWorkspaceContext",
        "summary": "Get workspace context",
        "description": "Returns the authenticated principal and trusted context for one accessible workspace.",
        "tags": [
          "Workspace context"
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "example": "019641a8-8c54-7f6c-8d2f-3fd1eb8b7531"
            },
            "required": true,
            "name": "workspaceId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Trusted workspace and principal context",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceContext"
                }
              }
            }
          },
          "401": {
            "description": "RFC 9457 Problem Details",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "RFC 9457 Problem Details",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "RFC 9457 Problem Details",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
...
  }
}